Move over UX code to Hover.tsx from QuickInfoSelectors.tsx
